// generated/discovery/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /** A `Datetime` scalar type represents a `Date` */
  Datetime: { input: any; output: any; }
  /** A `Hash` scalar type represents a `Record<string,any>` */
  Hash: { input: any; output: any; }
  /** A `RichText` scalar type represents a `string[]` (plainText or HTML) or a JSON */
  RichText: { input: any; output: any; }
};

export type BannerPiece = {
  __typename?: 'BannerPiece';
  banner: Maybe<Array<Maybe<Image>>>;
  callToAction: Maybe<CallToActionPiece>;
  description: Maybe<Scalars['RichText']['output']>;
  layout: Maybe<LayoutPiece>;
  title: Maybe<Scalars['String']['output']>;
};
// ...
export type CallToActionPiece = {
  __typename?: 'CallToActionPiece';
  action: Maybe<Array<Maybe<ContentChunksComponentForCallToActionPieceAction>>>;
};
// ...
export type ComponentChoiceComponentForLayoutPieceBackgroundMedia = {
  __typename?: 'ComponentChoiceComponentForLayoutPieceBackgroundMedia';
  image: Maybe<Array<Maybe<Image>>>;
  video: Maybe<Array<Maybe<VideoComponent>>>;
};
// ...
export type ContentChunksComponentForCallToActionPieceAction = {
  __typename?: 'ContentChunksComponentForCallToActionPieceAction';
  label: Maybe<Scalars['String']['output']>;
  url: Maybe<Scalars['String']['output']>;
};
// ...
export type Document = {
  /** The Aliases */
  aliases: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  /** The Breadcrumbs */
  breadcrumbs: Maybe<Array<Maybe<Array<Maybe<Breadcrumb>>>>>;
  /** The date when the version was created */
  createdAt: Maybe<Scalars['Datetime']['output']>;
  /** The canonical depth in the tree */
  depth: Maybe<Scalars['String']['output']>;
  /** The external reference */
  externalReference: Maybe<Scalars['String']['output']>;
  /** The History of Paths */
  history: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  /** The version id that includes the language and the publication state */
  id: Maybe<Scalars['ID']['output']>;
  /** The date when the document was indexed for the las time */
  indexedAt: Maybe<Scalars['Datetime']['output']>;
  /** The item id */
  itemId: Maybe<Scalars['ID']['output']>;
  /** The Language */
  language: Maybe<Scalars['String']['output']>;
  /** The name */
  name: Maybe<Scalars['String']['output']>;
  /** To be used with searchAfter and searchBefore */
  paginationToken: Maybe<Scalars['String']['output']>;
  /** The parent document */
  parent: Maybe<Document>;
  /** The parent id */
  parentId: Maybe<Scalars['ID']['output']>;
  /** The path */
  path: Maybe<Scalars['String']['output']>;
  /** The method to resolve the path */
  pathResolutionMethod: Maybe<PathResolutionMethod>;
  /** The canonical position in the tree */
  position: Maybe<Scalars['String']['output']>;
  /** The publication state */
  publicationState: Maybe<Scalars['String']['output']>;
  /** The date when the version was published */
  publishedAt: Maybe<Scalars['Datetime']['output']>;
  /** The score */
  score: Maybe<Scalars['Float']['output']>;
  /** The shape identifier */
  shape: Maybe<Scalars['String']['output']>;
  /** The Shortcuts */
  shortcuts: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  /** The different topic paths */
  topicPaths: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  /** The different topics */
  topics: Maybe<Scalars['Hash']['output']>;
  /** The type */
  type: Maybe<Scalars['String']['output']>;
  /** The date when the version was updated */
  updatedAt: Maybe<Scalars['Datetime']['output']>;
};
// ...
export type Image = {
  __typename?: 'Image';
  altText: Maybe<Scalars['String']['output']>;
  caption: Maybe<Scalars['RichText']['output']>;
  createdAt: Maybe<Scalars['Datetime']['output']>;
  focalPoint: Maybe<Scalars['Hash']['output']>;
  height: Maybe<Scalars['Int']['output']>;
  key: Maybe<Scalars['String']['output']>;
  meta: Maybe<Scalars['Hash']['output']>;
  /** Showcases that this image describes. */
  showcases: Maybe<Array<Maybe<Showcase>>>;
  topics: Maybe<Scalars['Hash']['output']>;
  updatedAt: Maybe<Scalars['Datetime']['output']>;
  url: Maybe<Scalars['String']['output']>;
  variants: Maybe<Array<Maybe<ImageVariant>>>;
  width: Maybe<Scalars['Int']['output']>;
};
// ...
export type ImageVariant = {
  __typename?: 'ImageVariant';
  height: Maybe<Scalars['Int']['output']>;
  key: Maybe<Scalars['String']['output']>;
  size: Maybe<Scalars['Int']['output']>;
  url: Maybe<Scalars['String']['output']>;
  width: Maybe<Scalars['Int']['output']>;
};
// ...
export type ItemRelationsInterface = {
  items: Maybe<Array<Maybe<Document>>>;
};
// ...
export type LayoutPiece = {
  __typename?: 'LayoutPiece';
  backgroundMedia: Maybe<ComponentChoiceComponentForLayoutPieceBackgroundMedia>;
  displayWidth: Maybe<Scalars['Hash']['output']>;
  theme: Maybe<Scalars['Hash']['output']>;
};
// ...
export type ProductVariantForProduct = Images & ProductVariant & {
  __typename?: 'ProductVariantForProduct';
  attributes: Maybe<Scalars['Hash']['output']>;
  /** This is a syntactic sugar for getting the 'default' priceVariant value in the priceVariants list. */
  defaultPrice: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'default' priceVariant. */
  defaultPriceFor: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'default' priceVariant. */
  defaultPriceForEveryone: Maybe<Scalars['Float']['output']>;
  /** This is a syntactic sugar for getting the 'default' stockLocation value in the stockLocations list. */
  defaultStock: Maybe<Scalars['Float']['output']>;
  description: Maybe<ComponentChoiceComponentForVariantProductDescription>;
  dimensions: Maybe<DimensionsPiece>;
  /** This is a syntactic sugar for getting the first image in the images list */
  firstImage: Maybe<Image>;
  images: Maybe<Array<Maybe<Image>>>;
  isDefault: Maybe<Scalars['Boolean']['output']>;
  matchingProducts: Maybe<ItemRelations>;
  /** This is a syntactic sugar for getting the 'membership-no' priceVariant value in the priceVariants list. */
  membershipNoPrice: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'membership-no' priceVariant. */
  membershipNoPriceFor: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'membership-no' priceVariant. */
  membershipNoPriceForEveryone: Maybe<Scalars['Float']['output']>;
  /** This is a syntactic sugar for getting the 'membership' priceVariant value in the priceVariants list. */
  membershipPrice: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'membership' priceVariant. */
  membershipPriceFor: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'membership' priceVariant. */
  membershipPriceForEveryone: Maybe<Scalars['Float']['output']>;
  /** The Subscription plan for monthly-fixed. */
  monthlyFixedSubscriptionPlan: Maybe<SubscriptionPlan>;
  name: Maybe<Scalars['String']['output']>;
  /** This is a syntactic sugar for getting the 'norway-b2b' priceVariant value in the priceVariants list. */
  norwayB2bPrice: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'norway-b2b' priceVariant. */
  norwayB2bPriceFor: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'norway-b2b' priceVariant. */
  norwayB2bPriceForEveryone: Maybe<Scalars['Float']['output']>;
  /** This is a syntactic sugar for getting the 'norway-b2c-sales' priceVariant value in the priceVariants list. */
  norwayB2cSalesPrice: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'norway-b2c-sales' priceVariant. */
  norwayB2cSalesPriceFor: Maybe<Scalars['Float']['output']>;
  /** PriceList aware best price for everyone for the 'norway-b2c-sales' priceVariant. */
  norwayB2cSalesPriceForEveryone: Maybe<Scalars['Float']['output']>;
  /** This is a syntactic sugar for getting the 'oslo' stockLocation value in the stockLocations list. */
  osloStock: Maybe<Scalars['Float']['output']>;
  priceVariants: Maybe<Scalars['Hash']['output']>;
  product: Maybe<Product>;
  sku: Maybe<Scalars['String']['output']>;
  stockLocations: Maybe<Scalars['Hash']['output']>;
  /** The Subscription plans for the variant. */
  subscriptionPlans: Maybe<Array<Maybe<SubscriptionPlan>>>;
  /** This is a syntactic sugar for getting the 'toronto' stockLocation value in the stockLocations list. */
  torontoStock: Maybe<Scalars['Float']['output']>;
  videos: Maybe<Array<Maybe<VideoComponent>>>;
};
// ...
export type Showcase = ItemRelationsInterface & SkuRelationsInterface & {
  __typename?: 'Showcase';
  hotspot: Maybe<Scalars['Hash']['output']>;
  items: Maybe<Array<Maybe<Document>>>;
  meta: Maybe<Scalars['Hash']['output']>;
  variants: Maybe<Array<Maybe<ProductVariant>>>;
};
// ...
export type SkuRelationsInterface = {
  variants: Maybe<Array<Maybe<ProductVariant>>>;
};
// ...
export type VideoComponent = {
  __typename?: 'VideoComponent';
  key: Maybe<Scalars['String']['output']>;
  playlists: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  thumbnails: Maybe<Array<Maybe<Image>>>;
  title: Maybe<Scalars['String']['output']>;
  url: Maybe<Scalars['String']['output']>;
};
// ...
export type BannerFragment = { __typename?: 'BannerPiece', title: string | null, description: any | null, layout: { __typename?: 'LayoutPiece', displayWidth: any | null, theme: any | null, backgroundMedia: { __typename?: 'ComponentChoiceComponentForLayoutPieceBackgroundMedia', image: Array<{ __typename?: 'Image', url: string | null, altText: string | null, focalPoint: any | null, variants: Array<{ __typename?: 'ImageVariant', url: string | null, width: number | null, height: number | null } | null> | null } | null> | null, video: Array<{ __typename?: 'VideoComponent', playlists: Array<string | null> | null, thumbnails: Array<{ __typename?: 'Image', url: string | null, altText: string | null, caption: any | null, meta: any | null, focalPoint: any | null, createdAt: any | null } | null> | null } | null> | null } | null } | null, banner: Array<{ __typename?: 'Image', url: string | null, altText: string | null, focalPoint: any | null, variants: Array<{ __typename?: 'ImageVariant', url: string | null, width: number | null, height: number | null } | null> | null, showcases: Array<{ __typename?: 'Showcase', hotspot: any | null, items: Array<{ __typename?: 'brand', name: string | null, path: string | null } | { __typename?: 'category', name: string | null, path: string | null } | { __typename?: 'landingPage', name: string | null, path: string | null } | { __typename?: 'menuItem', name: string | null, path: string | null } | { __typename?: 'product', name: string | null, path: string | null, defaultVariant: { __typename?: 'ProductVariantForProduct', defaultPrice: any | null, firstImage: { __typename?: 'Image', url: string | null, variants: Array<{ __typename?: 'ImageVariant', width: number | null, url: string | null, height: number | null } | null> | null } | null } | null } | { __typename?: 'story', name: string | null, path: string | null } | null> | null, variants: Array<{ __typename?: 'ProductVariantForProduct', name: string | null, defaultPrice: any | null, product: { __typename?: 'product', path: string | null } | null, firstImage: { __typename?: 'Image', url: string | null, variants: Array<{ __typename?: 'ImageVariant', width: number | null, url: string | null, height: number | null } | null> | null } | null } | null> | null } | null> | null } | null> | null, callToAction: { __typename?: 'CallToActionPiece', action: Array<{ __typename?: 'ContentChunksComponentForCallToActionPieceAction', label: string | null, url: string | null } | null> | null } | null };